r/host_virtual_switch: Change timeouts to 5 mins
Instead of using provider timeout * 0.8.

There's no guarantee that the above would have addressed *all* timeouts
in the intended fashion (basically the intention was to ensure that the
user could not set a timeout shorter than a hardcoded value). Having the
hardcoded timeout of 5 minutes is good enough for firing an error anyway
- if a host network operation is taking more than 5 minutes to
complete there is definitely a problem.