r/vmfs_disks: Work off of ScsiLun instead, filter and rescan
Using QueryAvailableDisksForVmfs only gives you the disks that are
unused and can be added to a new VMFS filesystem, which will create some
pretty unstable results and will pretty much guarantee that this needs
to be used with ignore_changes.

Instead, we now scan ScsiLun from the HostStorageSystem MO's properties
directly, and *only* work off of disks represented by the HostScsiDisk
type, which represents a actual fixed disk that can be used with VMFS.
Note that these nodes are still whole disks and do not represent
partitions, which needs to be taken into consideration when using this
data to create VMFS datastores.

Also added options for rescanning as this will more than likely be
needed as storage is provisioned to nodes (definitely need on nested
ESXi as new disk devices are not picked up until a rescan is done), and
a simple regular expression filter option to filter results based on an
expression of the canonical name, which should be helpful when trying to
make sure that only disks of a certain kind, controller, or target are
picked up by the data source and fed to something like
vsphere_vmfs_datastore.